requirements.txt
# requirements.txt
streamlit>=1.24.0
indic-nlp-library>=0.91
transformers>=4.30.2
torch>=2.2.0
tensorflow>=2.13.0
pandas>=2.0.3
numpy>=1.24.3
python-Levenshtein>=0.21.1
scikit-learn>=1.2.2